move collections-add to shared components
