+++ frontend/src/pages/org/index.ts
@@ -16,7 +16,6 @@ import "./workflows-list";
 import "./workflows-new";
 import "./crawl-detail";
 import "./crawls-list";
-import "./collections-add";
 import "./collections-list";
 import "./collections-new";
 import "./collection-edit";
